pyproject.toml
[project]
name = "p18-nano-banana-core-model"
version = "0.1.0"
description = ""
authors = [
    {name = "Antonin",email = "antonin@presti.ai"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.3.0",
    "pillow>=10.4.0",
    "numpy>=2.1.0",
    "streamlit>=1.38.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "tqdm>=4.66.0"
]

[tool.poetry]
packages = [{include = "p18_nano_banana_core_model", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
